requirements.txt
dnspython==2.7.0
mfrc522==0.0.7
paho-mqtt==2.1.0
pymongo==4.10.1
RPi.GPIO==0.7.1
spidev==3.6
